+++ src/react-query/feedbackList.ts
@@ -5,36 +5,65 @@
 import {
   InvalidateQueryFilters,
   QueryClient,
-  QueryFunctionContext,
-  QueryKey,
   useQuery,
   UseQueryResult,
   useSuspenseQuery,
   UseSuspenseQueryResult,
 } from "@tanstack/react-query";
-import { InkeepAnalyticsCore } from "../core.js";
-import { feedbackList } from "../funcs/feedbackList.js";
-import { combineSignals } from "../lib/primitives.js";
-import { RequestOptions } from "../lib/sdks.js";
-import * as components from "../models/components/index.js";
+import {
+  ConnectionError,
+  InvalidRequestError,
+  RequestAbortedError,
+  RequestTimeoutError,
+  UnexpectedClientError,
+} from "../models/errors/httpclienterrors.js";
+import * as errors from "../models/errors/index.js";
+import { InkeepAnalyticsError } from "../models/errors/inkeepanalyticserror.js";
+import { ResponseValidationError } from "../models/errors/responsevalidationerror.js";
+import { SDKValidationError } from "../models/errors/sdkvalidationerror.js";
 import * as operations from "../models/operations/index.js";
-import { unwrapAsync } from "../types/fp.js";
 import { useInkeepAnalyticsContext } from "./_context.js";
 import {
   QueryHookOptions,
   SuspenseQueryHookOptions,
   TupleToPrefixes,
 } from "./_types.js";
+import {
+  buildFeedbackListQuery,
+  FeedbackListQueryData,
+  prefetchFeedbackList,
+  queryKeyFeedbackList,
+} from "./feedbackList.core.js";
+export {
+  buildFeedbackListQuery,
+  type FeedbackListQueryData,
+  prefetchFeedbackList,
+  queryKeyFeedbackList,
+};
 
-export type FeedbackListQueryData = components.GetAllFeedbackResponse;
+export type FeedbackListQueryError =
+  | errors.BadRequest
+  | errors.Unauthorized
+  | errors.Forbidden
+  | errors.NotFound
+  | errors.UnprocessableEntity
+  | errors.InternalServerError
+  | InkeepAnalyticsError
+  | ResponseValidationError
+  | ConnectionError
+  | RequestAbortedError
+  | RequestTimeoutError
+  | InvalidRequestError
+  | UnexpectedClientError
+  | SDKValidationError;
 
 /**
  * Get All Feedback
  */
 export function useFeedbackList(
   request: operations.GetAllFeedbackRequest,
-  options?: QueryHookOptions<FeedbackListQueryData>,
-): UseQueryResult<FeedbackListQueryData, Error> {
+  options?: QueryHookOptions<FeedbackListQueryData, FeedbackListQueryError>,
+): UseQueryResult<FeedbackListQueryData, FeedbackListQueryError> {
   const client = useInkeepAnalyticsContext();
   return useQuery({
     ...buildFeedbackListQuery(
@@ -51,8 +80,11 @@ export function useFeedbackList(
  */
 export function useFeedbackListSuspense(
   request: operations.GetAllFeedbackRequest,
-  options?: SuspenseQueryHookOptions<FeedbackListQueryData>,
-): UseSuspenseQueryResult<FeedbackListQueryData, Error> {
+  options?: SuspenseQueryHookOptions<
+    FeedbackListQueryData,
+    FeedbackListQueryError
+  >,
+): UseSuspenseQueryResult<FeedbackListQueryData, FeedbackListQueryError> {
   const client = useInkeepAnalyticsContext();
   return useSuspenseQuery({
     ...buildFeedbackListQuery(
@@ -61,19 +93,6 @@ export function useFeedbackListSuspense(
       options,
     ),
     ...options,
-  });
-}
-
-export function prefetchFeedbackList(
-  queryClient: QueryClient,
-  client$: InkeepAnalyticsCore,
-  request: operations.GetAllFeedbackRequest,
-): Promise<void> {
-  return queryClient.prefetchQuery({
-    ...buildFeedbackListQuery(
-      client$,
-      request,
-    ),
   });
 }
 
@@ -116,44 +135,4 @@ export function invalidateAllFeedbackList(
     ...filters,
     queryKey: ["@inkeep/inkeep-analytics", "feedback", "list"],
   });
-}
-
-export function buildFeedbackListQuery(
-  client$: InkeepAnalyticsCore,
-  request: operations.GetAllFeedbackRequest,
-  options?: RequestOptions,
-): {
-  queryKey: QueryKey;
-  queryFn: (context: QueryFunctionContext) => Promise<FeedbackListQueryData>;
-} {
-  return {
-    queryKey: queryKeyFeedbackList({
-      limit: request.limit,
-      offset: request.offset,
-    }),
-    queryFn: async function feedbackListQueryFn(
-      ctx,
-    ): Promise<FeedbackListQueryData> {
-      const sig = combineSignals(ctx.signal, options?.fetchOptions?.signal);
-      const mergedOptions = {
-        ...options,
-        fetchOptions: { ...options?.fetchOptions, signal: sig },
-      };
-
-      return unwrapAsync(feedbackList(
-        client$,
-        request,
-        mergedOptions,
-      ));
-    },
-  };
-}
-
-export function queryKeyFeedbackList(
-  parameters: {
-    limit?: number | null | undefined;
-    offset?: number | null | undefined;
-  },
-): QueryKey {
-  return ["@inkeep/inkeep-analytics", "feedback", "list", parameters];
 }