+++ src/react-query/queryExportFeedbackQueryResults.ts
@@ -12,6 +12,17 @@ import { queryExportFeedbackQueryResults } from "../funcs/queryExportFeedbackQue
 import { combineSignals } from "../lib/primitives.js";
 import { RequestOptions } from "../lib/sdks.js";
 import * as components from "../models/components/index.js";
+import {
+  ConnectionError,
+  InvalidRequestError,
+  RequestAbortedError,
+  RequestTimeoutError,
+  UnexpectedClientError,
+} from "../models/errors/httpclienterrors.js";
+import * as errors from "../models/errors/index.js";
+import { InkeepAnalyticsError } from "../models/errors/inkeepanalyticserror.js";
+import { ResponseValidationError } from "../models/errors/responsevalidationerror.js";
+import { SDKValidationError } from "../models/errors/sdkvalidationerror.js";
 import { unwrapAsync } from "../types/fp.js";
 import { useInkeepAnalyticsContext } from "./_context.js";
 import { MutationHookOptions } from "./_types.js";
@@ -25,18 +36,33 @@ export type QueryExportFeedbackQueryResultsMutationData = ReadableStream<
   Uint8Array
 >;
 
+export type QueryExportFeedbackQueryResultsMutationError =
+  | errors.BadRequest
+  | errors.Unauthorized
+  | errors.Forbidden
+  | errors.UnprocessableEntity
+  | errors.InternalServerError
+  | InkeepAnalyticsError
+  | ResponseValidationError
+  | ConnectionError
+  | RequestAbortedError
+  | RequestTimeoutError
+  | InvalidRequestError
+  | UnexpectedClientError
+  | SDKValidationError;
+
 /**
  * Export Feedback Query Results
  */
 export function useQueryExportFeedbackQueryResultsMutation(
   options?: MutationHookOptions<
     QueryExportFeedbackQueryResultsMutationData,
-    Error,
+    QueryExportFeedbackQueryResultsMutationError,
     QueryExportFeedbackQueryResultsMutationVariables
   >,
 ): UseMutationResult<
   QueryExportFeedbackQueryResultsMutationData,
-  Error,
+  QueryExportFeedbackQueryResultsMutationError,
   QueryExportFeedbackQueryResultsMutationVariables
 > {
   const client = useInkeepAnalyticsContext();